+++ test/functional/auto_html_for_test.rb
@@ -6,7 +6,7 @@ class Article < ActiveRecord::Base
     html_escape
     youtube(:width => 400, :height => 250)
     image
-    link(:target => "_blank", :rel => "nofollow")
+    link(:target => "_blank")
     simple_format
   end
 end
@@ -35,6 +35,6 @@ class AutoHtmlForTest < Test::Unit::TestCase
     @article = Article.create!(:body => 'Yo!')
     @article.update_attributes(:body => 'http://vukajlija.com')
     @article.save!
-    assert_equal '<p><a href="http://vukajlija.com" rel="nofollow" target="_blank">http://vukajlija.com</a></p>', @article.body_html
+    assert_equal '<p><a href="http://vukajlija.com" target="_blank">http://vukajlija.com</a></p>', @article.body_html
   end
 end